+++ NEWS.md
@@ -1,7 +1,14 @@
-# NMdata 0.0.7
-Configuration mechanism improved
+# NMdata 0.0.6.6
+Central configuration mechanism implemented. The configuration
+function to use is NMdataConf. You can configure several options,
+corresponding to default values of arguments to different functions in
+the package. This is very useful if you want to change the directory
+and file naming structure, or if you want to change default column
+names.
 
-as.fun="data.table"
+The exclusion flag functions flagsAssign and flagsCount have been
+generalized to use customizable column names for the numerical and
+caracter flags. The default can be configured using NMdataConf.
 
 # NMdata 0.0.6.4
 If all common column names two data objects to merge are not used for